requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
bcrypt==5.0.0
billiard==4.2.2
build==1.3.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
chromadb==1.3.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
Django==4.2.26
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
durationpy==0.10
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-texttospeech==2.33.0
google-genai==1.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kubernetes==34.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pillow==11.3.0
posthog==5.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
Twisted==25.5.0
txaio==23.6.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
zope.interface==8.0.1
